requirements.txt
asgiref==3.4.1
boto3==1.18.20
botocore==1.21.20
dj-database-url==0.5.0
Django==3.2.5
django-allauth==0.41.0
django-crispy-forms==1.12.0
django-isbn-field==0.5.2
django-soft-delete==0.9.18
django-softdelete==0.9.2
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
oauthlib==3.1.1
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-stdnum==1.16
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
sqlparse==0.4.1
stripe==2.60.0
